Create MACD_Elder_Impulse_System.efs
Elder Impulse system rules

Inertia and Power
A good measure of the inertia of any trading vehicle is the slope of its fast EMA.
A rising EMA reflects bullish inertia, while a falling EMA reflects bearish inertia. The
power of any trend is reflected in the slope of MACD-Histogram.


1.	EMA rising & MACD-Histogram rising (especially below zero) = Impulse is green, bullish.
	Shorting prohibited, buying or standing aside permitted.
2.	EMA falling & MACD-Histogram falling (especially above zero) = Impulse is red, bearish.
	Buying prohibited, shorting or standing aside permitted.
3.	EMA rising & MACD-Histogram falling = Impulse is blue, neutral. Nothing is prohibited.
4.	EMA falling & MACD-Histogram rising = Impulse is blue, neutral. Nothing is prohibited.

The Impulse System is a censorship system. It doesn’t tell what to do—but it clearly tells what not to do.

Default Settings
Moving Average: 13, exponential.
MACD: 12,26,9